Backport r12743: [Ref] Silence Clang 10 warning about inexact result in implicit integral-to-floatingpoint conversions when the source value is std::numeric_limits::max().
git-svn-id: https://source.openmpt.org/svn/openmpt/branches/OpenMPT-1.27@13453 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/misc_util.h
@@ -466,11 +466,11 @@ inline Tdst saturate_cast(Tsrc src)
 template <typename Tdst>
 inline Tdst saturate_cast(double src)
 {
-	if(src >= std::numeric_limits<Tdst>::max())
+	if(src >= static_cast<double>(std::numeric_limits<Tdst>::max()))
 	{
 		return std::numeric_limits<Tdst>::max();
 	}
-	if(src <= std::numeric_limits<Tdst>::min())
+	if(src <= static_cast<double>(std::numeric_limits<Tdst>::min()))
 	{
 		return std::numeric_limits<Tdst>::min();
 	}
@@ -480,11 +480,11 @@ inline Tdst saturate_cast(double src)
 template <typename Tdst>
 inline Tdst saturate_cast(float src)
 {
-	if(src >= std::numeric_limits<Tdst>::max())
+	if(src >= static_cast<float>(std::numeric_limits<Tdst>::max()))
 	{
 		return std::numeric_limits<Tdst>::max();
 	}
-	if(src <= std::numeric_limits<Tdst>::min())
+	if(src <= static_cast<float>(std::numeric_limits<Tdst>::min()))
 	{
 		return std::numeric_limits<Tdst>::min();
 	}
